Create a more complex HTTP server using Express

+++ 0x08-making_change/0-making_change.py
@@ -12,10 +12,10 @@ def makeChange(coins, total):
     noc = 0
     coins.sort(reverse=True)
     for i in coins:
-        q = total / i
+        q = int(total / i)
         noc += q
         total %= i
 
     if total > 0:
         return -1
-    return int(noc)
+    return noc
